refactor: split up spade boolops module
